Add Animation Tests
